fixes #8
