+++ src/backend/project/api-project-op/README.md
@@ -1,0 +1,5 @@
+# OP模块说明
+
+OP （Operation） 是后台运营的操作类接口
+
+主要定义一些提供给运营后台系统调用的接口
