+++ src/backend/project/api-project-sample/src/main/kotlin/com/tencent/devops/project/api/UserProjectUserResource.kt
@@ -24,11 +24,40 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-dependencies {
-    compile project(":project:biz-project-op") // OP管理类接口
-    compile project(":project:biz-project-blueking") // 这是对接蓝鲸的实现
+package com.tencent.devops.project.api
 
-    // 根据依赖蓝鲸的打出来的包就是对接蓝鲸平台的微服务
+import com.tencent.devops.common.api.auth.AUTH_HEADER_DEVOPS_USER_ID
+import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID_DEFAULT_VALUE
+import com.tencent.devops.project.pojo.Result
+import com.tencent.devops.project.pojo.user.ProjectUser
+import com.tencent.devops.project.pojo.user.UserDeptDetail
+import io.swagger.annotations.Api
+import io.swagger.annotations.ApiOperation
+import io.swagger.annotations.ApiParam
+import javax.ws.rs.*
+import javax.ws.rs.core.MediaType
+
+@Api(tags = ["USER_PROJECT_USER"], description = "项目列表用户接口")
+@Path("/user/users")
+@Produces(MediaType.APPLICATION_JSON)
+@Consumes(MediaType.APPLICATION_JSON)
+interface UserProjectUserResource {
+
+    @GET
+    @Path("/")
+    @ApiOperation("查询用户基本信息")
+    fun get(
+        @ApiParam("用户ID", required = true, defaultValue = AUTH_HEADER_USER_ID_DEFAULT_VALUE)
+        @HeaderParam(AUTH_HEADER_DEVOPS_USER_ID)
+        userId: String
+    ): Result<ProjectUser>
+
+    @GET
+    @Path("/detail")
+    @ApiOperation("查询用户详细信息")
+    fun getDetail(
+        @ApiParam("userId", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_USER_ID)
+        userId: String
+    ): Result<UserDeptDetail>
 }
-
-apply from: "$rootDir/task_spring_boot_package.gradle"
