+++ src/backend/project/biz-project-blueking/build.gradle
@@ -26,6 +26,7 @@
 
 dependencies {
     compile project(":project:biz-project")
+    compile project(":project:api-project-blueking")
     compile project(":common:common-auth:common-auth-blueking")
 }
