+++ src/backend/settings.gradle
@@ -47,9 +47,13 @@ include ":common:common-auth:common-auth-blueking"
 
 include ":project"
 include ":project:api-project"
+include ":project:api-project-sample"
+include ":project:api-project-blueking"
+include ":project:api-project-op"
 include ":project:biz-project"
 include ":project:biz-project-blueking"
 include ":project:biz-project-sample"
+include ":project:biz-project-op"
 include ":project:boot-project"
 include ":project:model-project"
